Second pass refactor of fixture functions
