+++ acceptance_tests/features/environment.py
@@ -7,19 +7,17 @@ from structlog import wrap_logger
 from acceptance_tests import browser
 from acceptance_tests.features.fixtures import setup_data_survey_with_internal_user, \
     setup_data_with_2_enrolled_respondent_users_and_internal_user, \
+    setup_data_with_enrolled_respondent_user_and_collection_exercise_to_live, \
     setup_data_with_enrolled_respondent_user_and_internal_user, \
     setup_data_with_enrolled_respondent_user_and_internal_user_and_new_iac_and_collection_exercise_to_live, \
-    setup_with_internal_user, \
+    setup_data_with_internal_user, setup_data_with_internal_user_and_collection_exercise_to_created_status, \
     setup_data_with_internal_user_and_social_collection_exercise_to_closed_status, \
-    setup_data_with_internal_user_and_collection_exercise_to_created_status, \
-    setup_data_with_internal_user, \
-    setup_data_with_unenrolled_respondent_user, \
+    setup_data_with_respondent_user_data_and_new_iac, setup_data_with_unenrolled_respondent_user, \
     setup_data_with_unenrolled_respondent_user_and_internal_user, \
-    setup_data_with_unenrolled_respondent_user_and_new_iac, \
     setup_data_with_unenrolled_respondent_user_and_new_iac_and_collection_exercise_to_live, \
-    setup_survey_metadata_with_internal_user, setup_data_with_enrolled_respondent_user_and_collection_exercise_to_live
-
-from common import survey_utilities
+    setup_survey_metadata_with_internal_user, setup_with_internal_user
+from common.collection_exercise_utilities import execute_collection_exercises
+from common.respondent_utilities import register_respondent
 from config import Config
 from exceptions import MissingFixtureError
 
@@ -48,8 +46,8 @@ fixture_scenario_registry = {
         setup_data_with_unenrolled_respondent_user,
     'fixture.setup.data.with.unenrolled.respondent.user.and.internal.user':
         setup_data_with_unenrolled_respondent_user_and_internal_user,
-    'fixture.setup.data.with.unenrolled.respondent.user.and.new.iac':
-        setup_data_with_unenrolled_respondent_user_and_new_iac,
+    'fixture.setup.data.with.respondent.user.data.and.new.iac':
+        setup_data_with_respondent_user_data_and_new_iac,
     'fixture.setup.data.with.internal.user.and.collection.exercise.to.created.status':
         setup_data_with_internal_user_and_collection_exercise_to_created_status,
     'fixture.setup.data.with.internal.user.and.social.collection.exercise.to.closed.status':
@@ -66,10 +64,9 @@ fixture_scenario_registry = {
 
 
 def before_all(_):
-
     # Run all tests using original method - standalone tests run in sequence
     if not is_ignore_sequential_data_setup():
-        survey_utilities.setup_sequential_data_for_test()
+        setup_sequential_data_for_test()
 
 
 def before_feature(context, feature):
@@ -97,7 +94,7 @@ def before_scenario(context, scenario):
     logger.info(f'Running Feature [{context.feature_name}], Scenario [{context.scenario_name}]')
 
     # Default to non-standalone fixed user name, standalone mode changes it
-    context.respondent_user_name = Config.RESPONDENT_USERNAME
+    context.respondent_email = Config.RESPONDENT_USERNAME
     context.internal_user_name = Config.INTERNAL_USERNAME
 
     # Run any custom scenario setup from fixture tags
@@ -140,3 +137,9 @@ def get_survey_type(tags):
         return "Social"
 
     return 'Business'
+
+
+def setup_sequential_data_for_test():
+    execute_collection_exercises()
+    register_respondent(survey_id='cb8accda-6118-4d3b-85a3-149e28960c54', period='201801',
+                        username=Config.RESPONDENT_USERNAME, ru_ref=49900000001)
