+++ common/collection_exercise_utilities.py
@@ -4,9 +4,9 @@ from logging import getLogger
 
 from structlog import wrap_logger
 
-from common import common_utilities
-from controllers import collection_exercise_controller, sample_controller, database_controller, party_controller, \
-    case_controller
+import common.string_utilities
+from controllers import case_controller, collection_exercise_controller, database_controller, party_controller, \
+    sample_controller
 from controllers.collection_instrument_controller import get_collection_instruments_by_classifier, \
     link_collection_instrument_to_exercise, upload_seft_collection_instrument
 from controllers.database_controller import poll_collection_exercise_until_state_changed
@@ -125,7 +125,8 @@ def make_user_description(user_description_in, is_social_survey, max_field_lengt
     else:
         prefix = ''.join((USER_DESCRIPTION_BUSINESS_PREFIX, FIELD_SEPARATOR))
 
-    compacted_user_description = common_utilities.compact_string(user_description_in, max_field_length - len(prefix))
+    compacted_user_description = common.string_utilities.compact_string(user_description_in,
+                                                                        max_field_length - len(prefix))
 
     return ''.join((prefix, compacted_user_description))
 
@@ -142,3 +143,19 @@ def find_case_by_enrolment_code(enrolment_code):
 
 def get_party_from_email(email):
     return party_controller.get_party_by_email(email)
+
+
+def create_business_survey_period(period_offset_days=0):
+    period_date = datetime.utcnow() + timedelta(days=period_offset_days)
+
+    return format_period(period_date.year, period_date.month)
+
+
+def create_social_survey_period(period_offset_days=0):
+    period_date = datetime.utcnow() + timedelta(days=period_offset_days)
+
+    return format_period(period_date.year, period_date.month)
+
+
+def format_period(period_year, period_month):
+    return f'{period_year}{str(period_month).zfill(2)}'
