+++ common/common_utilities.py
@@ -10,20 +10,6 @@ def create_utc_timestamp():
     return datetime.utcnow().strftime('%Y-%m-%d-%H-%M-%S-%f')
 
 
-def compact_string(string_in, max_length):
-    string_out = string_in
-
-    # Retain as much of the string as possible, init cap all words, remove spaces and truncate if necessary
-    if len(string_in) > max_length:
-        string_out = string_out.title()
-        string_out = string_out.replace(' ', '')
-
-        if len(string_out) > max_length:
-            string_out = string_out[:max_length]
-
-    return string_out
-
-
 def create_behave_stop_argument(stop_on_failure):
     try:
         arg = strtobool(stop_on_failure)
