+++ common/internal_utilities.py
@@ -2,7 +2,7 @@ import json
 
 import requests
 
-from common import survey_utilities
+import common.respondent_utilities
 from config import Config
 
 
@@ -12,7 +12,7 @@ def create_internal_user_login_account(user_name):
     access_token = generate_access_token(client_id=Config.UAA_CLIENT_ID, client_secret=Config.UAA_SECRET, url=token_url)
 
     user_url = f'{Config.UAA_SERVICE_URL}/Users'
-    email = survey_utilities.make_email_address(user_name, user_name)
+    email = common.respondent_utilities.make_email_address(user_name, user_name)
 
     return create_user(access_token, user_name, Config.UAA_DEFAULT_USER_PASSWORD, email, user_name, user_name, user_url)
